requirements.txt
sentence-transformers
pinecone-client[grpc]
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
idna==3.10
numpy==2.2.4
pgvector==0.4.0
pydantic==2.11.2
pydantic_core==2.33.1
python-dotenv==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
structlog==25.2.0
typing-inspection==0.4.0
typing_extensions==4.13.1
uvicorn==0.34.0
